requirements.txt
Flask==2.3.2
gunicorn==21.2.0
pymongo==4.3.3
python-dotenv==1.0.0
requests==2.31.0
pytest==7.3.1
black==23.3.0
flake8==6.0.0
pydantic==2.0.2
SQLAlchemy==2.0.9
Werkzeug==2.3.6
Jinja2==3.1.2
Babel==2.12.1
PyJWT==2.7.0
pytest-cov==4.1.0 